refactor: Split mock_ddgs into separate files (1 class per file)
- Split mock_ddgs.py (3 classes) into 3 separate files:
  - mock_ddgs_search_client.py (MockDDGS)
  - mock_ddgs_context_manager.py (MockDDGSContextManager)
  - mock_ddgs_class.py (MockDDGSClass)
- Extract _convert_organic_results() in serper_client.py to reduce nesting

Now fully compliant with 1 class per file rule.

Remaining: services/content_scraper.py has 6-level nesting (needs extraction)
